+++ src/repull/models/get_airbnb_alteration_response_200.py
@@ -56,7 +56,7 @@ class GetAirbnbAlterationResponse200:
         field_dict.update(self.additional_properties)
         field_dict.update({
             "data": data,
-            "data_freshness": data_freshness,
+            "dataFreshness": data_freshness,
         })
 
         return field_dict
@@ -73,7 +73,7 @@ class GetAirbnbAlterationResponse200:
 
 
 
-        data_freshness = AirbnbDataFreshness.from_dict(d.pop("data_freshness"))
+        data_freshness = AirbnbDataFreshness.from_dict(d.pop("dataFreshness"))